\name{brain}
\docType{data}
\alias{brain}
\title{Brain image}
\description{
  This file contains data of a brain image. It has
  217x217 pixels. Gray levels are in the range [0, 255].
  In the data file, observations are listed as a 217
  x217 matrix. This image has blur involved.}
\usage{brain}
\format{This dataset is saved in an ascii file.}
