\name{circles}
\docType{data}
\alias{circles}
\title{Image of circles}
\description{
  This file contains data of the original circles image.
  It has 256x256 pixels. Gray levels are in the range
  [0, 255]. In the data file, observations are listed as
  a 256x256 matrix.}
\usage{circles}
\format{This dataset is saved in an ascii file.}
