\name{kid}
\docType{data}
\alias{kid}
\title{Image of a kid taking test}
\description{ This file contains data of original kid image.
  The image has 387x387 pixels. Gray levels are in the range
  [0, 255]. In the data file, observations are listed as
  a 387x387 matrix. This image has spatially variant blur
  involved.}
\usage{kid}
\format{This dataset is saved in an ascii file.}
\references{
Kang, Y., and Qiu, P., "Jump Detection in Blurred Regression 
Surfaces," \emph{Technometrics}, \bold{56}, 2014, 539-550.
}