\name{lena}
\docType{data}
\alias{lena}
\title{Image of Lena}
\description{
  This file contains data of the original Lena image.
  It has 512x512 pixels. Gray levels are in the
  range [0, 255]. In the data file, observations are
  listed as a 512x512 matrix.}
\usage{lena}
\format{This dataset is saved in an ascii file.}
\references{ November 1972 issue of Playboy magazine }