\name{peppers}
\docType{data}
\alias{peppers}
\title{Image of peppers}
\description{
  This file contains data of the original peppers image.
  It has 512x512 pixels. Gray levels are in the
  range [0, 255]. In the data file, observations are
  listed as a 512x512 matrix.}
\usage{peppers}
\format{This dataset is saved in an ascii file.}

