\name{sar}
\docType{data}
\alias{sar}
\title{Synthetic aperture radar image of an area near Thetford forest,
  England}
\description{ This file contains data of original sar image.
  The image has 250x250 pixels. Gray levels are in the range
  [0, 255]. In the data file, observations are listed as
  a 250x250 matrix. This image contains much noise.}
\usage{sar}
\format{This dataset is saved in an ascii file.}
\references{This data can be downloaded from:
  http://peipa.essex.ac.uk/ipa/pix/books/glasbey-horgan/}