\name{stopsign}
\docType{data}
\alias{stopsign}
\title{Image of Stop Sign}
\description{ This file contains data of stop sign image.
  The image has 160x160 pixels. Gray levels are in the range
  [0, 255]. In the data file, observations are listed as
  a 160x160 matrix. This image has much blurring involved.}
\usage{stopsign}
\format{This dataset is saved in an ascii file.}