\name{brain}
\docType{data}
\alias{brain}
\title{Brain Image}
\description{
    This file contains data of a brain image. It has
    217x217 pixels. Gray levels are in the range [0, 255].
    In the data file, observations are listed as a 217
    x217 matrix. This image has blur involved.}
\usage{brain}
\format{A matrix containing 217x217 pixels.}
\author{Yicheng Kang}
\references{
    Kang, Y., Mukherjee, P.S. and Qiu, P. (2018) ``Efficient Blind
    Image Deblurring Using Nonparametric Regression and Local
    Pixel Clustering'', \emph{Technometrics}, \bold{60}(4),
    522 -- 531, \doi{10.1080/00401706.2017.1415975}.
}
\keyword{datasets}