\name{circles}
\docType{data}
\alias{circles}
\title{Image of Circles}
\description{
    This file contains the circles image. It has 256x256 pixels.
    Gray levels are in the range [0, 255]. In the data file,
    observations are listed as a 256x256 matrix.}
\usage{circles}
\format{A matrix of 256x256 pixels.}
\author{Yicheng Kang}
\references{
    Qiu, P. (2005) \emph{Image Processing and Jump Regression
    Analysis}. New Jersey: Wiley.
}
\keyword{datasets}
