\name{kid}
\docType{data}
\alias{kid}
\title{Image of a Kid}
\description{
    This file contains the kid image. The image has 387x387
    pixels. Gray levels are in the range [0, 255]. In the data
    file, observations are listed as a 387x387 matrix. This image
    has spatially variant blur involved.}
\usage{kid}
\format{A matrix of 387x387 pixels.}
\author{Yicheng Kang}
\references{
    Kang, Y. and Qiu, P. (2014) "Jump Detection in Blurred 
    Regression Surfaces," \emph{Technometrics}, \bold{56}(4),
    539 -- 550, \doi{10.1080/00401706.2013.844732}.
}
\keyword{datasets}