\name{lena}
\docType{data}
\alias{lena}
\title{Lena Image}
\description{
  This file contains the Lena image. It has 512x512 pixels. Gray
  levels are in the range [0, 255]. In the data file, observations
  are listed as a 512x512 matrix.}
\usage{lena}
\format{A 512x512 matrix.}
\author{Yicheng Kang}
\source{ November 1972 issue of Playboy magazine.}
