\name{peppers}
\docType{data}
\alias{peppers}
\title{Image of Peppers}
\description{
  This file contains the peppers image. It has 512x512 pixels.
  Gray levels are in the range [0, 255]. In the data file,
  observations are  listed as a 512x512 matrix.}
\usage{peppers}
\format{A matrix of 512x512 pixels.}
\author{Yicheng Kang}
\references{
Kang, Y. (2020) ``Consistent Blind Image Deblurring Using
Jump-Preserving Extrapolation'', \emph{Journal of Computational and
 Graphical Statistics}, \bold{29}(2), 372 -- 382,
 \doi{10.1080/10618600.2019.1665536}.
}
\keyword{datasets}

