\name{sar}
\docType{data}
\alias{sar}
\title{Synthetic Aperture Radar Image}
\description{
    This file contains the synthetic aperture radar (SAR) image
    of an area near Thetford forest, England. The image has
    250x250 pixels. Gray levels are in the range [0, 255]. In the
    data file, observations are listed as a 250x250 matrix. This
    image is noisy.}
\usage{sar}
\format{A matrix of 250x250 pixels.}
\source{\url{https://users.phhp.ufl.edu/pqiu/research/book/data/index.html}}
\author{Yicheng Kang}
\references{
    Qiu, P. (2005) \emph{Image Processing and Jump Regression
    Analysis}. New Jersey: Wiley.
}
\keyword{datasets}