\name{stopsign}
\docType{data}
\alias{stopsign}
\title{Stop Sign Image}
\description{
    This file contains the stop sign image. The image has 160x160
    pixels. Gray levels are in the range [0, 255]. In the data
    file, observations are listed as a 160x160 matrix. This image
    has much blurring involved.}
\usage{stopsign}
\format{A matrix of 160x160 pixels.}
\author{Yicheng Kang}
\references{
Kang, Y. (2020) ``Consistent Blind Image Deblurring Using
Jump-Preserving Extrapolation'', \emph{Journal of Computational and
 Graphical Statistics}, \bold{29}(2), 372 -- 382,
 \doi{10.1080/10618600.2019.1665536}.
}
\keyword{datasets}